Refactor TypeRef out of value bindings and add type kinds
