+++ llvmlite/binding/__init__.py
@@ -11,6 +11,7 @@ from .passmanagers import *
 from .targets import *
 from .transforms import *
 from .value import *
+from .typeref import *
 from .analysis import *
 from .object_file import *
 from .context import *
