+++ llvmlite/binding/value.py
@@ -4,6 +4,7 @@ import enum
 
 from llvmlite.binding import ffi
 from llvmlite.binding.common import _decode_string, _encode_string
+from llvmlite.binding.typeref import TypeRef
 
 
 class Linkage(enum.IntEnum):
@@ -77,37 +78,6 @@ class ValueKind(enum.IntEnum):
 
     instruction = 24
     poison_value = 25
-
-
-class TypeRef(ffi.ObjectRef):
-    """A weak reference to a LLVM type
-    """
-    @property
-    def name(self):
-        """
-        Get type name
-        """
-        return ffi.ret_string(ffi.lib.LLVMPY_GetTypeName(self))
-
-    @property
-    def is_pointer(self):
-        """
-        Returns true is the type is a pointer type.
-        """
-        return ffi.lib.LLVMPY_TypeIsPointer(self)
-
-    @property
-    def element_type(self):
-        """
-        Returns the pointed-to type. When the type is not a pointer,
-        raises exception.
-        """
-        if not self.is_pointer:
-            raise ValueError("Type {} is not a pointer".format(self))
-        return TypeRef(ffi.lib.LLVMPY_GetElementType(self))
-
-    def __str__(self):
-        return ffi.ret_string(ffi.lib.LLVMPY_PrintType(self))
 
 
 class ValueRef(ffi.ObjectRef):
@@ -509,17 +479,6 @@ ffi.lib.LLVMPY_SetValueName.argtypes = [ffi.LLVMValueRef, c_char_p]
 
 ffi.lib.LLVMPY_TypeOf.argtypes = [ffi.LLVMValueRef]
 ffi.lib.LLVMPY_TypeOf.restype = ffi.LLVMTypeRef
-
-
-ffi.lib.LLVMPY_PrintType.argtypes = [ffi.LLVMTypeRef]
-ffi.lib.LLVMPY_PrintType.restype = c_void_p
-
-ffi.lib.LLVMPY_TypeIsPointer.argtypes = [ffi.LLVMTypeRef]
-ffi.lib.LLVMPY_TypeIsPointer.restype = c_bool
-
-ffi.lib.LLVMPY_GetElementType.argtypes = [ffi.LLVMTypeRef]
-ffi.lib.LLVMPY_GetElementType.restype = ffi.LLVMTypeRef
-
 
 ffi.lib.LLVMPY_GetTypeName.argtypes = [ffi.LLVMTypeRef]
 ffi.lib.LLVMPY_GetTypeName.restype = c_void_p
